requirements.txt
google-api-python-client==2.153.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
google-generativeai==0.8.3
Pillow==11.0.0
